requirements.txt
Flask
pymysql
gspread
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
oauth2client
mysql-connector-python
